+++ src/proto_018_Proxford/lib_delegate/state_transitions.ml
@@ -104,6 +104,15 @@ let prepare_preattest_action state proposal =
   in
   Prepare_preattestations {preattestations}
 
+let prepare_consensus_votes_action state proposal =
+  let preattestations : unsigned_consensus_vote_batch =
+    make_consensus_vote_batch state proposal Baking_state.Preattestation
+  in
+  let attestations : unsigned_consensus_vote_batch =
+    make_consensus_vote_batch state proposal Baking_state.Attestation
+  in
+  Prepare_consensus_votes {preattestations; attestations}
+
 let update_proposal ~is_proposal_applied state proposal =
   let open Lwt_syntax in
   let* () = Events.(emit updating_latest_proposal proposal.block.hash) in
@@ -150,6 +159,21 @@ let preattest state proposal =
     (* Here, we do not cancel pending signatures as it is already done
        in [handle_proposal]. *)
     return (new_state, prepare_preattest_action state proposal)
+
+let prepare_consensus_votes state proposal =
+  let open Lwt_syntax in
+  if Baking_state.is_first_block_in_protocol proposal then
+    (* We do not vote for the first transition block *)
+    let new_state = update_current_phase state Idle in
+    return (new_state, Do_nothing)
+  else
+    let* () = Events.(emit attempting_vote_proposal proposal.block.hash) in
+    let new_state =
+      (* We have detected a new proposal that needs to be voted for.
+         We switch to the `Awaiting_preattestations` phase. *)
+      update_current_phase state Awaiting_preattestations
+    in
+    return (new_state, prepare_consensus_votes_action state proposal)
 
 let extract_pqc state (new_proposal : proposal) =
   match new_proposal.block.prequorum with
@@ -213,14 +237,20 @@ let has_already_been_handled state new_proposal =
   Block_hash.(current_proposal.block.hash = new_proposal.block.hash)
   && state.level_state.is_latest_proposal_applied
 
+let mark_awaiting_pqc state =
+  let new_round_state =
+    {state.round_state with awaiting_unlocking_pqc = true}
+  in
+  let new_state = {state with round_state = new_round_state} in
+  new_state
+
 let rec handle_proposal ~is_proposal_applied state (new_proposal : proposal) =
   let open Lwt_syntax in
-  (* We need to avoid to send preattestations, if we are in phases were
-     preattestations have been sent already. This is needed to avoid switching
+  (* We need to avoid to send votes if we are in phases where consensus
+     votes are already being forged. This is needed to avoid switching
      back from Awaiting_attestations to Awaiting_preattestations.
-     Hypothesis: a fresh round's initial phase is Idle
-  *)
-  let may_preattest state proposal =
+     Hypothesis: a fresh round's initial phase is Idle *)
+  let may_vote state proposal =
     match state.round_state.current_phase with
     | Idle ->
         (* We prioritize the new and meaningful consensus vote signing
@@ -230,7 +260,7 @@ let rec handle_proposal ~is_proposal_applied state (new_proposal : proposal) =
            active forge requests are still processed until
            completion. *)
         state.global_state.forge_worker_hooks.cancel_all_pending_tasks () ;
-        preattest state proposal
+        prepare_consensus_votes state proposal
     | _ -> do_nothing state
   in
   let current_level = state.level_state.current_level in
@@ -299,6 +329,11 @@ let rec handle_proposal ~is_proposal_applied state (new_proposal : proposal) =
           let* new_state =
             may_update_proposal ~is_proposal_applied new_state new_proposal
           in
+          (* We invalidate early attestations *)
+          let new_round_state =
+            {new_state.round_state with early_attestations = []}
+          in
+          let new_state = {new_state with round_state = new_round_state} in
           (* The proposal is valid but maybe we already locked on a payload *)
           match new_state.level_state.locked_round with
           | Some locked_round -> (
@@ -307,27 +342,28 @@ let rec handle_proposal ~is_proposal_applied state (new_proposal : proposal) =
                   locked_round.payload_hash = new_proposal.block.payload_hash)
               then
                 (* when the new head has the same payload as our
-                   [locked_round], we accept it and preattest *)
-                may_preattest new_state new_proposal
+                   [locked_round], we accept it and vote for it *)
+                may_vote new_state new_proposal
               else
                 (* The payload is different *)
                 match new_proposal.block.prequorum with
                 | Some {round; _} when Round.(locked_round.round < round) ->
-                    (* This PQC is above our locked_round, we can preattest it *)
-                    may_preattest new_state new_proposal
+                    (* This PQC is above our locked_round, we can and vote for it *)
+                    may_vote new_state new_proposal
                 | _ ->
-                    (* We shouldn't preattest this proposal, but we
+                    (* We shouldn't vote for proposal, but we
                        should at least watch (pre)quorums events on it
                        but only when it is applied otherwise we await
                        for the proposal to be applied. *)
+                    let new_state = mark_awaiting_pqc new_state in
                     let new_state =
                       update_current_phase new_state Awaiting_preattestations
                     in
-                    return (new_state, Watch_proposal))
+                    return (new_state, Watch_prequorum))
           | None ->
               (* Otherwise, we did not lock on any payload, thus we can
-                 preattest it *)
-              may_preattest new_state new_proposal)
+                 vote for it it *)
+              may_vote new_state new_proposal)
   else
     (* Last case: new_proposal_level > current_level *)
     (* Possible scenarios:
@@ -339,7 +375,13 @@ let rec handle_proposal ~is_proposal_applied state (new_proposal : proposal) =
     let compute_new_state ~current_round ~delegate_slots
         ~next_level_delegate_slots =
       let round_state =
-        {current_round; current_phase = Idle; delayed_quorum = None}
+        {
+          current_round;
+          current_phase = Idle;
+          delayed_quorum = None;
+          early_attestations = [];
+          awaiting_unlocking_pqc = false;
+        }
       in
       let level_state =
         {
@@ -613,7 +655,13 @@ let end_of_round state current_round =
   (* We initialize the round's phase to Idle for the [handle_proposal]
      transition to trigger the preattestation action. *)
   let new_round_state =
-    {state.round_state with current_round = new_round; current_phase = Idle}
+    {
+      current_round = new_round;
+      current_phase = Idle;
+      delayed_quorum = None;
+      early_attestations = [];
+      awaiting_unlocking_pqc = false;
+    }
   in
   let new_state = {state with round_state = new_round_state} in
   (* we need to check if we need to bake for this round or not *)
@@ -703,6 +751,35 @@ let prepare_attest_action state proposal =
   in
   Prepare_attestations {attestations}
 
+let inject_early_arrived_attestations state =
+  let early_attestations = state.round_state.early_attestations in
+  match early_attestations with
+  | [] -> Lwt.return (state, Watch_quorum)
+  | first_signed_attestation :: _ as unbatched_signed_attestations -> (
+      let new_round_state = {state.round_state with early_attestations = []} in
+      let new_state = {state with round_state = new_round_state} in
+      let batch_branch = state.level_state.latest_proposal.predecessor.hash in
+      let batch_content =
+        let vote_consensus_content =
+          first_signed_attestation.unsigned_consensus_vote
+            .vote_consensus_content
+        in
+        {
+          level = vote_consensus_content.level;
+          round = vote_consensus_content.round;
+          block_payload_hash = vote_consensus_content.block_payload_hash;
+        }
+      in
+      make_signed_consensus_vote_batch
+        Attestation
+        batch_content
+        ~batch_branch
+        unbatched_signed_attestations
+      |> function
+      | Ok signed_attestations ->
+          Lwt.return (new_state, Inject_attestations {signed_attestations})
+      | Error _err -> (* Unreachable *) do_nothing new_state)
+
 let prequorum_reached_when_awaiting_preattestations state candidate
     preattestations =
   let open Lwt_syntax in
@@ -752,7 +829,14 @@ let prequorum_reached_when_awaiting_preattestations state candidate
         latest_proposal.block.payload_hash
     in
     let new_state = update_current_phase new_state Awaiting_attestations in
-    return (new_state, prepare_attest_action new_state latest_proposal)
+    if new_state.round_state.awaiting_unlocking_pqc then
+      (* We were locked and did not trigger preemptive
+         consensus votes: we need to start attesting now. *)
+      return (new_state, prepare_attest_action new_state latest_proposal)
+    else
+      (* We already triggered preemptive attestation forging, we
+         either have those already or we are waiting for them. *)
+      inject_early_arrived_attestations new_state
 
 let quorum_reached_when_waiting_attestations state candidate attestation_qc =
   let open Lwt_syntax in
@@ -840,6 +924,57 @@ let handle_expected_applied_proposal (state : Baking_state.t) =
   let new_state = update_current_phase new_state Idle in
   do_nothing new_state
 
+let handle_arriving_attestation state signed_attestation =
+  let open Lwt_syntax in
+  let {
+    vote_consensus_content =
+      {
+        level;
+        round = att_round;
+        block_payload_hash = att_payload_hash;
+        slot = _;
+      };
+    delegate;
+    _;
+  } =
+    signed_attestation.unsigned_consensus_vote
+  in
+  let att_level = Raw_level.to_int32 level in
+  let attestation_matches_proposal =
+    let {payload_hash; round = block_round; shell; _} =
+      state.level_state.latest_proposal.block
+    in
+    Block_payload_hash.(att_payload_hash = payload_hash)
+    && Round.(att_round = block_round)
+    && Compare.Int32.(shell.level = att_level)
+  in
+  if not attestation_matches_proposal then
+    let* () =
+      Events.(emit discarding_attestation (delegate, att_level, att_round))
+    in
+    do_nothing state
+  else
+    match state.round_state.current_phase with
+    | Awaiting_preattestations ->
+        (* An attestation is ready for injection but the prequorum has
+           not been reached yet: we save them until then. *)
+        let new_round_state =
+          {
+            state.round_state with
+            early_attestations =
+              signed_attestation :: state.round_state.early_attestations;
+          }
+        in
+        let new_state = {state with round_state = new_round_state} in
+        do_nothing new_state
+    | Idle | Awaiting_attestations | Awaiting_application ->
+        (* For these three phases, we have necessarily already reached the
+           prequorum: we are safe to inject. *)
+        let signed_attestations =
+          make_singleton_consensus_vote_batch signed_attestation
+        in
+        Lwt.return (state, Inject_attestations {signed_attestations})
+
 let handle_forge_event state forge_event =
   match forge_event with
   | Block_ready prepared_block ->
@@ -850,7 +985,7 @@ let handle_forge_event state forge_event =
   | Preattestation_ready signed_preattestation ->
       Lwt.return (state, Inject_preattestation {signed_preattestation})
   | Attestation_ready signed_attestation ->
-      Lwt.return (state, Inject_attestation {signed_attestation})
+      handle_arriving_attestation state signed_attestation
 
 (* Hypothesis:
    - The state is not to be modified outside this module